Add files via upload
The game has been created for my homework. First, size of the Pacifier board is requested. The player starts the game at the corner (1,1). The ghost starts at right corner (N,N). In each turn is asked to the user to move (up, down, right, left, up left, top-right, bottom-left, bottom-right). The program checks the movements is valid, if it goes outside(limits of the matrix), the user is warned and the movement requested again. The program calculate next movement of the ghost (it goes the closes point to the user). In this section distance between 2 points method is used. You can see the every movement in the screen and you can try to run away from the ghost.